Remove unopened comment close tag
Noticed a dangling close tag on the contact us page. Pull request is to remove the close, but perhaps the intent was to comment out the pateron link? 

https://www.crisiscleanup.org/contact

Crisis Cleanup LLC, a New Jersey LLC, is the agent for this implementation of the Crisis Cleanup project.

Please donate through our Patreon Page.

-->
Thank you for your time and interest.